requirements.txt
streamlit
torch
torchvision
torchinfo
pillow
matplotlib